refactor(SimpleBSTTests): change names of SimpleBSTForTest and TestSimpleBST classes
